changed some line of codes
